# feat: Call ConvertToSidecarScope function with multiple goroutines