**Please provide a description of this PR:**

issue #41453 

Call the `ConvertToSidecarScope` function with multiple goroutines to reduce the time consumption of the `initSidecarScopes` function.