@wulianglongrd 2022-10-19
hello @howardjohn  @ramaraochavali , I have submitted the benchmark code. Below are my benchmark results on my mac.

**test env:**
```code
goos: darwin
goarch: amd64
cpu: Intel(R) Core(TM) i5-7360U CPU @ 2.30GHz
```

**test result:**
Note that I changed the name of the "name" column. the `old` represent the original code, the `new` represent the code with multiple goroutines. Please read the benchmark source code for details.
```code
$ benchstat old.txt new.txt
name                                      old time/op    new time/op    delta
InitSidecarScopesThrottleService10-4       294µs ±20%     266µs ±17%   -9.32%  (p=0.004 n=20+20)
InitSidecarScopesThrottleService30-4       924µs ±14%     725µs ± 9%  -21.47%  (p=0.000 n=20+20)
InitSidecarScopesThrottleService50-4      1.97ms ±14%    1.41ms ±17%  -28.48%  (p=0.000 n=20+17)
InitSidecarScopesThrottleService100-4     5.67ms ±16%    3.77ms ±14%  -33.55%  (p=0.000 n=20+19)
InitSidecarScopesThrottleService200-4     19.1ms ±16%    12.0ms ±14%  -37.22%  (p=0.000 n=20+19)
InitSidecarScopesThrottleService300-4     45.3ms ±37%    27.6ms ±22%  -39.14%  (p=0.000 n=18+19)
InitSidecarScopesThrottleService400-4      110ms ±44%      50ms ±25%  -54.30%  (p=0.000 n=18+20)
InitSidecarScopesThrottleService500-4      111ms ±25%      78ms ±16%  -29.31%  (p=0.000 n=20+20)
InitSidecarScopesThrottleService1000-4     424ms ±40%     299ms ±22%  -29.52%  (p=0.000 n=19+20)

name                                      old alloc/op   new alloc/op   delta
InitSidecarScopesThrottleService10-4      74.0kB ± 0%    77.2kB ± 0%   +4.39%  (p=0.000 n=20+20)
InitSidecarScopesThrottleService30-4       229kB ± 0%     238kB ± 0%   +3.94%  (p=0.000 n=20+20)
InitSidecarScopesThrottleService50-4       407kB ± 0%     422kB ± 0%   +3.63%  (p=0.000 n=20+19)
InitSidecarScopesThrottleService100-4      943kB ± 0%     972kB ± 0%   +3.10%  (p=0.000 n=20+20)
InitSidecarScopesThrottleService200-4     2.38MB ± 0%    2.44MB ± 0%   +2.43%  (p=0.000 n=19+20)
InitSidecarScopesThrottleService300-4     4.40MB ± 0%    4.49MB ± 0%   +1.97%  (p=0.000 n=20+20)
InitSidecarScopesThrottleService400-4     6.61MB ± 0%    6.73MB ± 0%   +1.75%  (p=0.000 n=19+20)
InitSidecarScopesThrottleService500-4     9.46MB ± 0%    9.61MB ± 0%   +1.53%  (p=0.000 n=19+19)
InitSidecarScopesThrottleService1000-4    31.3MB ± 0%    31.6MB ± 0%   +0.92%  (p=0.000 n=20+20)

name                                      old allocs/op  new allocs/op  delta
InitSidecarScopesThrottleService10-4         757 ± 0%       793 ± 0%   +4.76%  (p=0.000 n=20+20)
InitSidecarScopesThrottleService30-4       2.84k ± 0%     2.94k ± 0%   +3.38%  (p=0.000 n=20+20)
InitSidecarScopesThrottleService50-4       5.72k ± 0%     5.88k ± 0%   +2.73%  (p=0.000 n=14+16)
InitSidecarScopesThrottleService100-4      16.4k ± 0%     16.7k ± 0%   +1.86%  (p=0.000 n=20+20)
InitSidecarScopesThrottleService200-4      52.8k ± 0%     53.4k ± 0%   +1.15%  (p=0.000 n=19+20)
InitSidecarScopesThrottleService300-4       109k ± 0%      110k ± 0%   +0.83%  (p=0.000 n=20+20)
InitSidecarScopesThrottleService400-4       186k ± 0%      187k ± 0%   +0.65%  (p=0.000 n=20+20)
InitSidecarScopesThrottleService500-4       282k ± 0%      284k ± 0%   +0.54%  (p=0.000 n=19+20)
InitSidecarScopesThrottleService1000-4     1.06M ± 0%     1.07M ± 0%   +0.28%  (p=0.000 n=20+20)
```

**benchmark cpu profile flamegraph**:
- before(the original code):

<img width="1318" alt="image" src="https://user-images.githubusercontent.com/14371864/196695734-4244eece-874e-4ba1-8e6a-a8ae44fc7ce2.png">

- after(with multiple goroutines):

<img width="1325" alt="image" src="https://user-images.githubusercontent.com/14371864/196695849-17e5d427-cac8-4721-b581-9a832fc6d142.png">

From the above test results, using multiple goroutines can reduce the time consumption of the `initSidecarScopes` function, and the impact on CPU and memory is not obvious.

thanks.