@howardjohn 2022-10-19
Its not true that the CPu usage is unchanged. You are looking at only relative CPU usage of different functions, if you look at the total CPU time spent its far greater with more goroutines.

Which makes sense - its  CPU bound process, its not possible to use less CPU by just changing which CPU runs the process..