@hzxuzhonghu 2022-10-20
@wulianglongrd Could you give a graph of istiod, we want to see how `ConvertToSidecarScope` impacts the whole system.

I can understand all you want is to speed up the calc, and decrease xds latency, irrespective of cpu/memory cost.  It makes sense in your case, if this function is the bottle, I don't mind providing such option to satisfy different cases. We need a complete trace graph first.