@wulianglongrd 2022-10-21
hello @howardjohn @hzxuzhonghu, Below are my perf results.

**perf process:**
1. Do not set `PILOT_SIDECARSCOPE_CONVERT_THROTTLE` to start and perf, get old.svg
2. Kill pilot-discovery
3. Set `export PILOT_SIDECARSCOPE_CONVERT_THROTTLE=4` to start and perf, get new.svg


**perf commands:**
```shell
$ perf record -F 99 -p {pid} -g -- sleep 40
$ perf script > old.perf
$ git clone https://github.com/brendangregg/FlameGraph
$ ./FlameGraph/stackcollapse-perf.pl old.perf > old.folded
$ ./FlameGraph/flamegraph.pl old.folded > old.svg
```

**perf result:**
- old.svg
![old](https://user-images.githubusercontent.com/14371864/197130412-ae5d91b1-a0dc-4c42-88bb-09b7f8ba82c1.svg)

- new.svg
![new](https://user-images.githubusercontent.com/14371864/197130462-57585f00-ddf1-4936-b389-dd93d7ec7bb6.svg)


thanks.