@wulianglongrd 2022-10-21
Additional instructions:
After pilot-discovery is fully started, during perf, I create a `dr`.