@wulianglongrd 2022-10-21
@hzxuzhonghu `old.trace.out` and `new.trace.out` all in the gz file.

[trace.tar.gz](https://github.com/istio/istio/files/9836540/trace.tar.gz)