@howardjohn 2022-10-21
The flamegraph doesn't really give any new or useful info IMO (its actually largely identical before and after, as well). The concern is that its

* Complex
* Not safe (CPU spikes are already problematic, this makes them far worse)
* Not fixing the right problem (do less work instead of splitting the work across CPUs)