@sschepens 2022-10-24
@wulianglongrd in the flamegraphs you posted it would seem your case is mostly governed by `MostSpecificHostMatch` which should theoretically be fast (a single map lookup in best case scenario), but apparently you're not hitting that case and instead fullscanning hostnames to match the most specific.

Could this maybe be an insight to look on?

@howardjohn i was looking into `MostSpecificHostMatch` wouldn't there be a more efficient way to find matching destination rules? maybe using a Trie structure instead of a list of destination rules, another way would be to actually separate wildcarded destinationrules so not to iterate all desintation rules until we find a matching one. Currently we're iterating all destination rules for every service needing one which doesn't have a equal destination rule.